[Bug #7910] - MPR and Fusion do not account for varied Modality LUTs in source images [Bug #9266] - Unable to render images with very small rescale slope (<<1)  * Added a normalization LUT concept to the greyscale image display pipeline.  * Implemented a normalization LUT for PET images with very small rescale slope (less than 1/(2^BitsStored-1)).  * Fixed a number of places where LUT implementations were unnecessarily rounding input/output range values prior to final output.

+++ ImageViewer/Imaging/LUTFactory.cs
@@ -435,8 +435,8 @@ namespace ClearCanvas.ImageViewer.Imaging
 		private class CachedModalityLutLinear : SimpleDataModalityLut
 		{
 			public CachedModalityLutLinear(IDataModalityLut source)
-				: base((int) source.MinInputValue, source.Data,
-						(int) source.MinOutputValue, (int) source.MaxOutputValue,
+				: base(source.MinInputValue, source.Data,
+						source.MinOutputValue, source.MaxOutputValue,
 						source.GetKey(), source.GetDescription())
 			{
 			}
@@ -493,13 +493,13 @@ namespace ClearCanvas.ImageViewer.Imaging
 
 			public override int MinInputValue
 			{
-				get { return (int) RealLut.MinInputValue; }
+				get { return RealLut.MinInputValue; }
 				set { }
 			}
 
 			public override int MaxInputValue
 			{
-				get { return (int) RealLut.MaxInputValue; }
+				get { return RealLut.MaxInputValue; }
 				set { }
 			}
